Rev-jump working.
Things we added:

Prior distributions for u, rn.
jx(u) = 0 far from p=0, and also jx(u -> -infty) < 1, to allow move to
the regular MCMC procedure.  This is important because we defined the
jump map to be deterministic, and it may get stuck.  Having jx no
saturate at 1 lets the sampler to not dwell too long if it gets stuck.